Define a host's "public suffix" and "registrable domain"
This patch attempts to formalize the way in which the Public Suffix List
relates to URLs, thereby making it possible to more easily explain the
limitations of features like `document.domain`, and cookies somewhat
absurd-seeming behavior across diverse hosts.

This patch does not expose an API to query this concept, as requested in
[1]. Doing so in a future patch should be quite straightforward,
however, if we decide that that's a reasonable thing to do.

[1]: https://www.w3.org/Bugs/Public/show_bug.cgi?id=25865