Add URLSearchParams.prototype.sort()
This method is added to increase cache hits when making requests. It’s
opt-in as the order of code points in a URL’s query is significant by
default. It’s up to applications to decide if name order is not
significant for them.

Tests: https://github.com/w3c/web-platform-tests/pull/4531.

Fixes #26.