Record whether the URL parser removed newlines.
In order to support checks like those sketched out in [1], it would
be helpful to record at parse-time whether or not tabs and newline
characters were stripped from a given URL. This patch does so by
adding a cleverly-named flag that could be referenced from other
specifications. Fetch, for instance.