Percent-encode additional characters in "fragment state".
Currently, we percent-encode characters in "fragment state" using the C0
control percent-encode set. Firefox encodes more than that, and it seems
reasonable to align around that behavior for reasons spelled out in #291
and the comments of #344.

This patch adds a new "fragment percent-encode set" which contains the
C0 control percent-encode set, along with:

* 0x22 (")
* 0x3C (<)
* 0x3E (>)
* 0x60 (`)

Closes #344.