Change query state slightly to better deal with non-UTF-8 encodings
If the input to the URL parser contains code points outside the non-UTF-8 encoding's value space and the URL parser was invoked using a non-UTF-8 encoding, then those code points end up as &#...;.

The problem is that &, #, and ; are also URL separators, but the previous algorithm would only encode #. This ensures that & and ; are also encoded, as some browsers already do, but only if they came about as the result of the encode operation.

Tests: https://github.com/w3c/web-platform-tests/pull/10915.

Fixes https://github.com/whatwg/encoding/issues/139.