Verify domain is not empty after Unicode ToASCII
The empty host is not allowed in special non-"file" URLs, but Unicode ToASCII can result in the empty string, even if the input is not empty, for example:

* the input consists entirely of IDNA ignored code points
* the input is `xn--`.

Domain to ASCII (which invokes Unicode ToASCII) does not fail when it is called from the host parser with such an input, because VerifyDnsLength is false. So it needs an additional check.

Tests: https://github.com/web-platform-tests/wpt/pull/23432.