Editorial: Add note about when ToASCII = ASCII lowercase
Many implementations currently skip ToASCII if domain is ASCII-only, but
as discovered in [1] and [2], this can result in some undesirable
behavior. Adding a note prevents implementors from making the mistake of
thinking ToASCII is a no-op if the input is ASCII, and also provides a
recommendation on how to properly optimize the ToASCII step.

[1]: https://github.com/whatwg/url/issues/267
[2]: https://github.com/whatwg/url/pull/309#issuecomment-301405332