Editorial: add note about when ToASCII equals ASCII lowercase
Many implementations currently skip ToASCII if domain is ASCII-only, but as discovered in https://github.com/whatwg/url/issues/267 and https://github.com/whatwg/url/pull/309#issuecomment-301405332, this can result in some undesirable behavior. Adding a note prevents implementers from making the mistake of thinking ToASCII is a no-op if the input is ASCII, and also provides a recommendation on how to properly optimize the ToASCII step.